+++ app/Models/UserPermission.php
@@ -24,6 +24,7 @@ class UserPermission extends Model
 
     public const USER_MANAGEMENT_EDIT_PERMISSION = 'userManagementEdit';
 
+    protected $keyType = 'string';
     public $incrementing = false;
 
     protected $fillable = [