+++ resources/views/auth/oauth-user-not-found.blade.php
@@ -1,0 +1,6 @@
+<x-guest-layout>
+    <div class="text-sm text-gray-800 text-center">
+        <p>{{ __('No account was found.')}}</p>
+        <p>{{ __('Please contact a club member to create an account before logging in.') }}</p>
+    </div>
+</x-guest-layout>